Set playback rate to 1x when livestream has reached live edge
- When buffer event happens, get player timeline window and compare
  playback position with default (close to edge) position. If current
  playback position is newer than default, set playback rate to 1x.
- Use setPlaybackSpeed() instead of setPlaybackParameters().

Note that this _can_ cause playback rate to reset if skipping to an area
that not loaded yet _and_ in front of the live edge. It shouldn't be a
major problem as when just watching at 2x to catch up on a livestream it
wouldn't go past the buffered area.