Chromecast support
Create PlayerManager and CastHelper classes to handle cast playback as
well as split code for FOSS variant.

PlayerManger initializes players and has methods for changing current
player and initializing player with media URL.

CastHelper holds the CastContext and handles cast button visibility and
the Chromecast introductory overlay.

Remove references to Players in MainActivity and FileViewFragment.
Players are always accessed through the PlayerManager instance.

MediaRouteButton is present in the layout files in the main source set,
since it is not part of a closed source library. Its visibility remains
GONE the entire time in the FOSS variant.

Fix: #148