claim context menu: add Report action, implement handlers on all screens
Currently, although all context menu actions are shown when opening the
context menu on claims on any screen, for most screens the actions don't
do anything. Implement the functionality for all screens.

In ClaimListAdapter, remove the longClickForContextMenu flag and always
set currentPosition to the item's absolute position in the adapter when
a long press occurs. In the context menu listeners, use
getCurrentPosition() instead of getPosition() as the latter doesn't
return the correct value on a long press.

On the search page, check if a claim is a channel, and if so don't
attempt to get its signing channel when blocking/muting.